rename component and style
