+++ src/widgets/index.js
@@ -1,0 +1,10 @@
+import ActiveCall from './webphone/ActiveCall/ActiveCall.react';
+
+const webphone = {
+  ActiveCall,
+};
+
+export {
+  webphone,
+  ActiveCall,
+};
