+++ src/widgets/webphone/ActiveCallWithNote/index.css
