+++ src/widgets/webphone/CallFooter/index.css
