+++ src/widgets/webphone/CallFooter/index.js
