+++ src/widgets/webphone/CallInfo/index.css
