+++ src/widgets/webphone/CallInfo/index.js
