+++ src/widgets/webphone/Closable/index.css
