+++ src/widgets/webphone/Closable/index.js
