+++ src/widgets/webphone/DropdownItem/index.css
