+++ src/widgets/webphone/Flip/index.css
