+++ src/widgets/webphone/Flip/index.js
