+++ src/widgets/webphone/Transfer/index.js
