+++ src/widgets/webphone/WebPhone/index.css
